+++ crypto/dsa/build.info
@@ -5,7 +5,7 @@ $COMMON=dsa_sign.c dsa_vrf.c dsa_lib.c dsa_ossl.c dsa_check.c \
 
 SOURCE[../../libcrypto]=$COMMON\
         dsa_asn1.c dsa_err.c dsa_ameth.c dsa_pmeth.c dsa_prn.c \
-        dsa_meth.c
+        dsa_meth.c dsa_ctrl.c
 IF[{- !$disabled{'deprecated-0.9.8'} -}]
   SOURCE[../../libcrypto]=dsa_depr.c
 ENDIF
