+++ crypto/dsa/dsa_lib.c
@@ -13,15 +13,11 @@
  */
 #include "internal/deprecated.h"
 
-#include <stdio.h>
-#include "internal/cryptlib.h"
-#include "internal/refcount.h"
-#include <openssl/bn.h>
-#include <openssl/asn1.h>
 #include <openssl/engine.h>
-#include "dsa_local.h"
 #include "crypto/dsa.h"
 #include "crypto/dh.h" /* required by DSA_dup_DH() */
+#include "internal/cryptlib.h"
+#include "dsa_local.h"
 
 static DSA *dsa_new_intern(ENGINE *engine, OSSL_LIB_CTX *libctx);
